+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- 
+- Removed the implementation of *AddAssign* for point. Point should be immutable.
 
 ## [0.3.0] - 2024-05-18